Wrote functions 4 & 5 (getAvgsOfStudents & writeFinalGrades).
IMPORTANT: The filepath in writeFinalGrades is specific to my computer and WILL NOT WORK on your computers. Be sure to change it before you test things yourself, or it will not work.

Since neither of the other functions have been written, I cannot easily test these functions yet. (I could probably write all of this project relatively quickly, but I'm leaving the other functions to y'all). I can help and make changes as necessary, so feel free to ask me anything about this.